requirements.txt
requests
beautifulsoup4
openai
python-dotenv
IPython
gradio
ollama
pillow

torch
huggingface_hub
transformers
diffusers
datasets
soundfile

bitsandbytes
accelerate